backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
django-extensions==3.2.0
django-extra-fields==3.0.2
django-filter==22.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
Django==4.1
djangorestframework-simplejwt==4.3.0
djangorestframework==3.13.1
djoser==2.1.0
drf-spectacular==0.23.1
flake8==5.0.4
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests-oauthlib==1.3.1
requests==2.28.1
setuptools==64.0.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
wheel==0.37.1
python-dotenv~=0.20.0
gunicorn==20.1.0
